refactor: move resolve_product_groups to file, improve doc, code structure
