+++ csaf-rs/src/csaf/traits/util/mod.rs
@@ -1,3 +1,4 @@
 pub mod extract_references;
 pub mod generic_with;
 pub mod not_present_20;
+pub mod resolve_product_groups;
